Code of Winning Methods/U3/requirements.txt
numpy==1.18.5
pandas==0.23.4
lightgbm==2.2.3
tensorflow-gpu==1.13.1
Keras==2.2.4
scipy==1.4.1
scikit-learn==0.23.1
tqdm==4.31.1
psutil==5.6.2


